+++ Plugin/shared_plugin/bukkit/src/main/java/com/kishku7/chunksmith/platform/BukkitConfig.java
@@ -107,6 +107,11 @@ public final class BukkitConfig implements Config {
     }
 
     @Override
+    public boolean isLodDhOverrideEnabled() {
+        return false;
+    }
+
+    @Override
     public void reload() {
         plugin.reloadConfig();
     }